refactor tests/test_sky_fovs.py to make a CLI tool for the tester
