+++ tests/README.txt
@@ -1,0 +1,7 @@
+You can run the automated test suite with this command (from the repository root dir)::
+
+    # run all tests
+    pytest
+
+    # skip slow tests
+    pytest -m "not slow"
