+++ tetra3/cli/generate_database.py
@@ -107,3 +107,6 @@ def main():
         epoch_proper_motion=args.epoch_proper_motion,
         pattern_stars_per_fov=args.pattern_stars_per_fov,
     )
+
+if __name__ == "__main__":
+    main()
